First Commit - Racing Post Data Scraper
This scraper requires you to input the URL that contains the list of horses on sale for a certain sale. It then scrapes all the horses and their relatives data. Roughly it takes 8 hours to run per sale. Each sale will contain over 1,000,000 data points that are scraped and stored across 9 different sheets within the 1 excel file. This file can then be analysed to see what mare pedigrees should be examined closer first.